moved `lmo.l_rv_{}` into `lmo.contrib.scipy_stats`
